pause, added separate fragment for adding a new deck

+++ app/src/main/java/com/albeartwo/mtgdeckmaker/database/DeckCardCrossRef.kt
@@ -9,7 +9,7 @@ import androidx.room.Entity
 
 data class DeckCardCrossRef(
 
-    val deck_db_id : Long ,
+    val deck_db_id : Int ,
     val card_db_id : Long ,
     val oracle_id : String ,
 )
